+++ dkcgbc.c
@@ -263,7 +263,7 @@ int dkcgbc_compress (struct COMPRESSOR *gbc) {
 
     /* happy defaults! */
     for (i = 0; i <= gbc->in.length; i++) {
-        static const struct PATH p = { NULL, -1llu, {0,0,0} };
+        static const struct PATH p = { NULL, (size_t)-1, {0,0,0} };
         steps[i] = p;
     }
     steps[0].used = 0;
